-Se agrego opcion de subir imagenes sin redmencionar. -Ya no se genera la carpeta file.-Se puede pasar una ruta destino personalizada

+++ components/uploader/Uploader.php
@@ -6,10 +6,11 @@ use yii\base\Component;
 use yii\web\UploadedFile;
 use yii\imagine\Image;
 
+
 /**
  * Uploader
  * Componente Yii para control de carga de archivos en servidor
- * @version 1.01 Febrero 2024 for PHP 8.0 > 8.2
+ * @version 1.01 16 de Febrero 2024 for PHP 8.0 > 8.2
  * @author Benjamin Castillo Arriaga <benjamin.castillo.arriaga@gmail.com>
  * 
  *  Usted puede consulta la libreria original del autor en: https://github.com/benjamin-castillo/yii-components/
@@ -23,7 +24,7 @@ class Uploader extends Component
     /**
      *  Define la carpeta donde se guardaran 
      */
-    const FILE_DIR = '/files';
+    //const FILE_DIR = '/files';
     const THUMB_DIR = '/thumbnails';
 
     public $errors;
@@ -89,38 +90,48 @@ class Uploader extends Component
     public function save($post)
     {
 
+        $this->basePath ? '' : die("<br>Ruta base ausente o vacia");
+
         $oFile = UploadedFile::getInstanceByName($post);
         // echo "<pre>";
         // var_dump($oFile);
         // echo "</pre>";
 
         if (empty($_FILES)) {
-            $this->errors .= "No se recibió el archivo";
+            $this->errors .= "No se recibió el archivo desde el POST";
             die("Error, no se recibió el archivo en el POST.");
             return;
         } else {
 
-            if (empty($this->fileName)) {
+            if (empty($this->fileName)) { //Si no se existe nombre personalizado
                 $this->fileName = $this->preserveBaseName ? $oFile->getBaseName() : date('dmy') . time() . rand();
             }
 
             // Obtiene nombre de archivo thumb
             $this->fileThumbName = $this->thumbnail ? $this->fileName . '_thumb' : NULL;
+
             //Obtiene nombre de extension
             $this->fileExtension = $oFile->getExtension();
 
             // Generar carpeta para archivo original
-            $this->filePath = $this->generatePath(self::FILE_DIR);
-            //echo "<br>Carpeta para archivo original:" . $this->filePath;
+            //$this->filePath = $this->generatePath(self::FILE_DIR);
+            //$this->logDebug .= "<br>Carpeta para archivo original:" . $this->filePath;
 
-            // Generar carpeta para imagen miniatura
-            $this->thumbnailPath = $this->generatePath(self::THUMB_DIR);
-            //echo "<br>Ruta minuatura:" . $this->thumbnailPath;
+
+            if ($this->thumbnail === true) { // Si desean generar miniatura
+                // Generar carpeta para imagen miniatura
+                $this->thumbnailPath = $this->generatePath(self::THUMB_DIR);
+                $this->logDebug .= "<br>Ruta minuatura:" . $this->thumbnailPath;
+            }
+
             $this->fileRoute = $this->filePath . '/' . $this->fileName . '.' . $this->fileExtension;
-            //echo "<br>Ruta de archivo:" . $this->fileRoute;
+            $this->logDebug .= "<br>fileRoute:" . $this->fileRoute;
+            // Agrega la ruta thumb
             $this->thumbRoute = $this->thumbnailPath . '/' . $this->fileThumbName . '.' . $this->fileExtension;
-            $this->fileType = $oFile->type;
+            $this->logDebug .= "<br>Ruta thumb:" . $this->thumbRoute;
 
+            // Setea extensión de archivo
+            $this->fileType = $oFile->type;
             $this->logDebug .= "<br>Tipo de archivo recibido:" . $oFile->type . "";
 
             if (strpos($oFile->type, 'image') !== FALSE) { //Si el tipo es cualquier tipo de imágen
@@ -145,8 +156,7 @@ class Uploader extends Component
      */
     public function saveFile($filePost)
     {
-        $tempPath = $this->getUploadPath() . $this->generatePath(self::TEMP_DIR);
-
+        $tempPath = $this->getUploadPath();
         $tempImg = $tempPath . '/' . $this->fileName . '.' . $this->fileExtension;
 
         $this->logDebug .= "<br>Archivo a guardar:" . $tempImg;
@@ -161,39 +171,57 @@ class Uploader extends Component
 
 
     /**
-     *  Procesa la imágen para redimensional
+     *  Procesa de manera opcional las imágenes para redimensionarlas y bajar la calidad
+     * Probado con los formatos png
+     * 
      * @param type $filePost
      */
     public function saveImage($filePost)
     {
-        $tempPath = $this->getUploadPath() . $this->generatePath(self::TEMP_DIR);
-        $tempImg = $tempPath . '/' . $this->fileName . '.' . $this->fileExtension;
-        if ($filePost->saveAs($tempImg)) { //Crea una copia del archivo original
-            //Obtiene tamaño de imagen
-            $tempImgSize = $this->getImgSize($tempImg);
-            //Redimeciona tamaño de imagen
-            $newImgSize = $this->getNewImageSize($tempImgSize['width'], $tempImgSize['height']);
 
-            $image = new Image(); // Crear un objeto de tipo imagen
-            // redimensiona la imagen
-            $image->getImagine()
-                ->open($tempImg)
-                ->thumbnail(new \Imagine\Image\Box($newImgSize['width'], $newImgSize['height']))
-                ->save($this->getUploadPath() . $this->fileRoute, ['quality' => 100]);
+        if ($this->thumbnail === true) { // Si desean generar miniatura
+            $tempPath = $this->getUploadPath() . $this->generatePath(self::TEMP_DIR);
+            $tempImg = $tempPath . '/' . $this->fileName . '.' . $this->fileExtension;
+            if ($filePost->saveAs($tempImg)) { //Crea una copia del archivo original
+                //Obtiene tamaño de imagen
+                $tempImgSize = $this->getImgSize($tempImg);
+                //Redimeciona tamaño de imagen
+                $newImgSize = $this->getNewImageSize($tempImgSize['width'], $tempImgSize['height']);
 
-            if ($this->thumbnail) {
-                $newImgThumbSize = $this->getNewImageSize($tempImgSize['width'], $tempImgSize['height'], TRUE);
-                $imageThumb = new Image();
-                $imageThumb->getImagine()
+                $image = new Image(); // Crear un objeto de tipo imagen
+                // redimensiona la imagen
+                $image->getImagine()
                     ->open($tempImg)
-                    ->thumbnail(new \Imagine\Image\Box($newImgThumbSize['width'], $newImgThumbSize['height']))
-                    ->save($this->getUploadPath() . $this->thumbRoute, ['quality' => 100]);
-            }
+                    ->thumbnail(new \Imagine\Image\Box($newImgSize['width'], $newImgSize['height']))
+                    ->save($this->getUploadPath() . $this->fileRoute, ['quality' => 100]);
 
-            if (!$this->preserveOriginalFile) {
-                unlink($tempImg);
+                if ($this->thumbnail) {
+                    $newImgThumbSize = $this->getNewImageSize($tempImgSize['width'], $tempImgSize['height'], TRUE);
+                    $imageThumb = new Image();
+                    $imageThumb->getImagine()
+                        ->open($tempImg)
+                        ->thumbnail(new \Imagine\Image\Box($newImgThumbSize['width'], $newImgThumbSize['height']))
+                        ->save($this->getUploadPath() . $this->thumbRoute, ['quality' => 100]);
+                }
+
+                if (!$this->preserveOriginalFile) {
+                    unlink($tempImg);
+                }
+            }
+        } else {// Si no se desea procesar la imagen
+            $tempPath = $this->getUploadPath();
+            $tempImg = $tempPath . '/' . $this->fileName . '.' . $this->fileExtension;
+
+            $this->logDebug .= "<br>Archivo a guardar:" . $tempImg;
+
+
+            if ($filePost->saveAs($tempImg)) { //Crea una el archivo recibido por post en el servidor
+                $this->logDebug .= "<br>Archivo cargado con exito en:" . $tempImg;
+            } else {
+                die("<br>Error no se logro guardar el archivo en el servidor web");
             }
         }
+
     }
 
     /**
@@ -262,15 +290,16 @@ class Uploader extends Component
     }
 
     /**
-     * Generar carpetas 
-     * @param type $path
+     * Genera carpeta en base a la ruta que se pasa como parámetro
+     * @param string $path // Debe ser la ruta absoluta
      * @return null
      */
     public function generatePath(string $path)
     {
         $pathTemp = $this->getUploadPath() . $path;
         if (!file_exists($pathTemp)) {
-            if (!mkdir($pathTemp, 0755, true)) {
+            //if (!mkdir($pathTemp, 0755, true)) {
+            if (!mkdir($pathTemp, 0777, true)) { // Cambio de tipo de permiso version beta
                 die("error al crear " . $pathTemp);
                 return NULL;
             }
@@ -300,7 +329,7 @@ class Uploader extends Component
 
     /**
      * Ingresar el nombre del archivo. No debe llevar nombre de extension
-     * @param type $path
+     * @param string $path
      * @return null
      */
     public function setFileName(string $filename)
@@ -316,6 +345,22 @@ class Uploader extends Component
     public function getFileName(string $filename)
     {
         return $this->fileName;
+    }
+
+    /**
+     * Ontiene información importánte del ambiente
+     * @param void
+     * @return array
+     */
+    public function getInfoEnviroment()
+    {
+        $enviroment = [
+            "upload_max_filesize" => ini_get('upload_max_filesize'),
+            "post_max_size" => ini_get('post_max_size'),
+            "memory_limit" => ini_get('memory_limit'),
+            "max_execution_time" => ini_get('max_execution_time')
+        ];
+        return $enviroment;
     }
